requirements.txt
Django>=1.11.16
Pillow==3.1.0
wheel==0.24.0
django-pagedown==1.0.6
django-markdown-deux
django-crispy-forms
